+++ Shared/UtilsLIB/UtilsLib/Utils.cpp
@@ -7,7 +7,7 @@ bool utils::compareTwoDoubles(const double& a, const double& b)
   return std::abs(a - b) < .00001;
 }
 
-bool utils::checkWithin(size_t real, size_t close, size_t diff)
+bool utils::checkWithin(int real, int close, int diff)
 {
   int realInt = static_cast<int>(real);
   return realInt >= static_cast<int>(close - diff) &&
